requirements.txt
python-dotenv
streamlit
langchain
langchain-openai
openai>=1.0.0
pypdf
python-docx
langchain-experimental
python-pptx
faiss-cpu
langchain-community
ruff
ragas